src/python_tools/requirements.txt
jupyter==1.0.0
notebook==6.4.1
numpy==1.19.4
matplotlib==3.3.3
protobuf==3.6.1
bokeh==2.2.3
ipywidgets==7.6.2
jupytext==1.11.2
tftpy==0.8.1
pandas==1.2.1
